core: Move `atomic` into a new `sync` module
This mirrors the same hierarchy in the standard library.
