+++ src/libcore/sync/atomic.rs
